Paints only the waveforms that are viewable
Fixes the waveform component to paint only the waveforms that are
viewable and to not paint the waveforms of signals that are in collapsed
groups.

Splits TreeNode into two classes: DirectoryNode and InspectedNode. This
is done because multiple copies of the same signal can be displayed at
the same time but requires an additional nodeId field to distinuish
between the nodes. Meanwhile, DirectoryNodes should be unique.

Adds a modified depth first search iterator that only traverses through
expanded nodes. The iterator is used in the waveform component for
painting.