Vertically align signal component and wave component
Signal component and wave component changed to share the same scroll bar
component.

Improve selecting and right click on signal component (clear selection if
click happens outside of a node, right click changes selection to the node
containing the click if that node is not originally selected).

Waveforms are now vertically centered.