Add support for combined signals (eg ready-valid)
Changes Transition to have a generic value in order to allow combined
signals to have transition values that are arrays. Waveforms and Signals
are then also forced to have a type parameter.

Redefines a waveform as simply a sequence of transitions (no name). A
signal is anything that has a waveform. There are 2 types of signals:
pure signals (with a waveform obtained from a VCD or treadle) and
modified signals (that are derived from pure signals). Within modified
signals, right now there's truncated signals (splitting a pure signal
such as a bundle) and combined signals (2+ pure signals combined to more
easily show semantic information). A pure signal has a name which refers
to the original name of the signal from the VCD or treadle.

WaveDisplaySettings in DisplayModel maps InspectionNodes to a
WaveDisplaySetting, rather than a Signal to a WaveDisplaySetting, which
allows two copies of the same signal to be drawn with different settings
at the same time.