Add ability to show signal dependency information
Change implementation of clock to use Treadle's ClockInfo.

Separate the main window from the treadle controller for better
isolation.

Add initializing parameter to Util.toValueChange method to make it easy
to append to an existing array of transitions.

Fix bug with the SignalNameRenderer where the value of a ready valid
combined signal is null.

Fix bug with painters where the sliding(2) method returns an iterator
with a single Array element that itself also contains a single element
if the original iterator only has 1 element to begin with.
E.g. Array(3).sliding(2).next() returns Array(3)